Analyze Direct Camera Pictures
React-Camera takes Images as a Base64 String (DataURL). It needs to be converted into ArrayBuffer to work with Azure API.

Also, some Online Image-URLs are also in Base64-DataURL. Now, we are coverting them to ArrayBuffer as well.

In the UI, the Camera Button opens the Camera Screen.
And the Upload Buttob opens the Upload/URL Component.

The Image Analysis is returned as it is.